LPS-159357 Move the new select to new file
